working project Shruti

+++ Backend/PatientService/src/main/java/com/vanammesis/patientservice/responses/PatientDetailsResponse.java
@@ -17,7 +17,7 @@ public class PatientDetailsResponse {
 
     private long patientDetailId;
 
-    private int patientId;
+    private long patientId;
 
     private List<String> patientSymptoms;
